requirements.txt
anthropic==0.42.0
fastapi==0.115.6
pydantic==2.10.4
python-dotenv==1.0.1
Requests==2.32.3
